Split Pattern classes into separate files
