+++ include/PatternView.h
@@ -1,0 +1,99 @@
+/*
+ * PatternView.h
+ *
+ * Copyright (c) 2004-2014 Tobias Doerffel <tobydox/at/users.sourceforge.net>
+ *
+ * This file is part of LMMS - https://lmms.io
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation; either
+ * version 2 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * License along with this program (see COPYING); if not, write to the
+ * Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA 02110-1301 USA.
+ *
+ */
+ 
+#ifndef PATTERN_VIEW_H
+#define PATTERN_VIEW_H
+ 
+#include "Pattern.h"
+#include "TrackContentObjectView.h"
+
+class Pattern;
+ 
+ 
+class PatternView : public TrackContentObjectView
+{
+	Q_OBJECT
+
+public:
+	PatternView( Pattern* pattern, TrackView* parent );
+ 	virtual ~PatternView() = default;
+
+	Q_PROPERTY(QColor noteFillColor READ getNoteFillColor WRITE setNoteFillColor)
+	Q_PROPERTY(QColor noteBorderColor READ getNoteBorderColor WRITE setNoteBorderColor)
+	Q_PROPERTY(QColor mutedNoteFillColor READ getMutedNoteFillColor WRITE setMutedNoteFillColor)
+	Q_PROPERTY(QColor mutedNoteBorderColor READ getMutedNoteBorderColor WRITE setMutedNoteBorderColor)
+
+	QColor const & getNoteFillColor() const { return m_noteFillColor; }
+	void setNoteFillColor(QColor const & color) { m_noteFillColor = color; }
+
+	QColor const & getNoteBorderColor() const { return m_noteBorderColor; }
+	void setNoteBorderColor(QColor const & color) { m_noteBorderColor = color; }
+
+	QColor const & getMutedNoteFillColor() const { return m_mutedNoteFillColor; }
+	void setMutedNoteFillColor(QColor const & color) { m_mutedNoteFillColor = color; }
+
+	QColor const & getMutedNoteBorderColor() const { return m_mutedNoteBorderColor; }
+	void setMutedNoteBorderColor(QColor const & color) { m_mutedNoteBorderColor = color; }
+
+public slots:
+	Pattern* getPattern();
+	void update() override;
+
+
+protected slots:
+	void openInPianoRoll();
+	void setGhostInPianoRoll();
+
+	void resetName();
+	void changeName();
+
+
+protected:
+	void constructContextMenu( QMenu * ) override;
+	void mousePressEvent( QMouseEvent * _me ) override;
+	void mouseDoubleClickEvent( QMouseEvent * _me ) override;
+	void paintEvent( QPaintEvent * pe ) override;
+	void wheelEvent( QWheelEvent * _we ) override;
+
+
+private:
+	static QPixmap * s_stepBtnOn0;
+	static QPixmap * s_stepBtnOn200;
+	static QPixmap * s_stepBtnOff;
+	static QPixmap * s_stepBtnOffLight;
+
+	Pattern* m_pat;
+	QPixmap m_paintPixmap;
+
+	QColor m_noteFillColor;
+	QColor m_noteBorderColor;
+	QColor m_mutedNoteFillColor;
+	QColor m_mutedNoteBorderColor;
+
+	QStaticText m_staticTextName;
+} ;
+ 
+ 
+ 
+ #endif
