+++ src/gui/CMakeLists.txt
@@ -28,6 +28,7 @@ SET(LMMS_SRCS
 	gui/MidiCCRackView.cpp
 	gui/MidiSetupWidget.cpp
 	gui/ModelView.cpp
+	gui/PatternView.cpp
 	gui/PeakControllerDialog.cpp
 	gui/PianoView.cpp
 	gui/PluginBrowser.cpp
